backend/lnd_sidecar/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
httpx==0.25.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0
aiofiles==23.2.1
apscheduler==3.10.4
# Azure packages (optional - uncomment if using Azure integration)
# azure-storage-blob==12.19.0
# msal==1.25.0

